Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/child-src/frame-allowed-expected.txt

ALERT: PASS
CONSOLE MESSAGE: line 21: IFrame load event fired: the IFrame's location is 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-pass.html'.
Frames should be governed by 'child-src'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The IFrame should not have been blocked. It wasn't.
PASS successfullyParsed is true

TEST COMPLETE
